refactor: extract JSON parsing and SSRF security into separate modules
- Extract planner JSON parsing/rescue logic (~274 lines) into
  core/planner_json.py for independent testing and reduced cognitive load
- Extract SSRF/DNS rebinding security functions (~214 lines) into
  tools/web_search_security.py to isolate security-critical code
- planner.py: 1,407 → 1,133 lines (-19.5%)
- web_search.py: 1,306 → 1,120 lines (-14.2%)
- All 107 web_search tests pass with full backward compatibility
- Update LARGE_FILE_REVIEW.md with completed improvements
